Added enhancement to process prices and dividends (#86)
* Added dividends to StockProcessor

* Added error upload step for process-stock

* Fixed error upload step

* Added enhancement to process prices and dividends

+++ equicast_ingestion/processor/stock.py
@@ -25,17 +25,21 @@ class StockProcessor:
 
     @staticmethod
     def _process_prices(extractor: StockDataExtractor, folder: str):
-        print(f"💲 Fetching stock prices for {extractor.ticker}.")
-        price_data = extractor.extract_stock_price_data()
-        if price_data:
-            price_data.to_parquet(os.path.join(folder, "stock_price.parquet"))
+        filename = os.path.join(folder, f"stock_price.parquet")
+        if not os.path.exists(filename):
+            print(f"💲 Fetching stock prices for {extractor.ticker}.")
+            price_data = extractor.extract_stock_price_data()
+            if price_data:
+                price_data.to_parquet(filename)
 
     @staticmethod
     def _process_dividends(extractor: StockDataExtractor, folder: str):
-        print(f"💰 Fetching dividends for {extractor.ticker}.")
-        dividends = extractor.extract_dividends()
-        if dividends:
-            dividends.to_parquet(os.path.join(folder, "dividends.parquet"))
+        filename = os.path.join(folder, f"dividends.parquet")
+        if not os.path.exists(filename):
+            print(f"💰 Fetching dividends for {extractor.ticker}.")
+            dividends = extractor.extract_dividends()
+            if dividends:
+                dividends.to_parquet(os.path.join(folder, "dividends.parquet"))
 
     def _process_ticker(self, ticker: str):
         print(f"📥 Fetching ticker data for {ticker}.")
